pyproject.toml
[project]
name = "notification"
version = "0.1.0"
description = ""
authors = [
    {name = "Denys",email = "dxth63@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]

[tool.poetry]
name = "reminder-api"
version = "0.1.0"
description = "A reminder API backend for app"
authors = ["Denys Harkavenko <dxth63@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
alembic = "^1.12.1"
email-validator = "^2.1.0.post1"

[tool.poetry.group.dev.dependencies]
pandas = "^2.2.3"
pytest-cov = "^6.1.1"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80            # Enforce 80-char lines.
indent-width = 4
target-version = "py313"    # Trager python version is 3.13
fix = true

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
 
[tool.ruff.lint]
# Enable all ruff rules by default, then ignore overlaping ones
select = ["ALL"]
extend-ignore =  [
    "D",
    "N",    # PEP8 naming handed by Pylint
    "PL",   # All Pylint-oequivalent rules must be handed by Pylint
    "E501", # Line-too-long exeption is handed by Pylint
    "D203",
]
fixable = ["ALL"]   # Allow fix for all enabled rules when `--fix` is provided
 
[tool.ruff.format]
quote-style = "double"  # Use double quotes for strings.
indent-style = "space"  # Like Black, indent with spaces, rather than tabs.
skip-magic-trailing-comma = false   # Respect magic trailing commas.
line-ending = "auto"    # Automatically detect the appropriate line ending.
 
[tool.ruff.lint.mccabe]
max-complexity = 5  # Cylcomatic complexity treshold by Thomas J. McCabe,
# for more information see https://en.wikipedia.org/wiki/Cyclomatic_complexity
[tool.pylint.format]
max-line-length = 80

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = [
    "F401", 
    "S101",
    "SLF001",
]

[tool.coverage.report]
exclude_also = [
    'if self.debug:',
    'if settings.DEBUG',
    'if __name__ == .__main__.:',
    'if TYPE_CHECKING:',

    'raise AssertionError',
    'raise NotImplementedError',
    '@(abc\.)?abstractmethod',
]
[tool.coverage.run]
omit = [
    "*/errors/*",
    "*/tests/*",
]

[tool.poetry.scripts]
start = "uvicorn src.main:app --reload"